+++ lib/ex_ghpr/cli/search.ex
@@ -1,0 +1,4 @@
+use Croma
+
+defmodule ExGHPR.CLI.Search do
+end
